Consolidate Status layout + Overview page into shared web-common components
- `StatusLayout.svelte`: side nav (Overview / Resources / Tables)
  parameterized by `basePath`. Both apps' status `+layout.svelte` thin-
  wrap it.
- `StatusOverviewPage.svelte`: project info card + resources +
  errors sections. Local passes `runtimeVersion` (via local service)
  and slots `TablesSection`; cloud uses defaults.
